test: Move some functions to a spec helpers file
